[CLI] Add a "restart" app command (#37)
The CLI provides a command to start and stop an app/example. Often it's useful to have a restart command, that will stop and start again the same app. This is needed only for the CLI interface.

The command already exists. But error management should be improved:

restarting an app X when app Y is running does not show a clear error
if the app running is not the one being restarted: error
if no app is running: just start the app